backend/requirements.txt
fastapi>=0.109.2
uvicorn[standard]>=0.27.1
strawberry-graphql[fastapi]>=0.219.0
sqlalchemy[asyncio]>=2.0.25
alembic>=1.14.1
asyncpg>=0.29.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
structlog>=24.1.0
pydantic>=2.6.1
redis>=5.0.1
